Added keras optimizer and reorganized modules.
Cleaned up the optimize class by modularizing pytorch and keras modules.

Added keras sequential model creation with config.

Added proper callback functionality to keras and pytorch models.
